Refactor AO, fix issue with transparents, add inset leaf model

+++ src/common.hpp
@@ -17,7 +17,7 @@ namespace cppcraft
 	const float BIOME_SCALE = 0.001;
   const float UNDERGEN_SCALE = 0.001;
 
-	typedef uint16_t light_value_t;
+	typedef uint32_t light_value_t;
 }
 
 extern void dump_trace();
